Framework_Core/core/executor.py
import pytest
import allure
from typing import List, Dict, Any
from loguru import logger

class TestExecutor:
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.setup_logging()

    def setup_logging(self):
        """配置日志"""
        logger.add("logs/test_execution.log", rotation="500 MB", level="INFO")

    @staticmethod
    def run_tests(test_paths: List[str], markers: List[str] = None) -> bool:
        """
        执行测试用例
        :param test_paths: 测试用例路径列表
        :param markers: pytest标记列表
        :return: 测试执行是否成功
        """
        try:
            args = ["-v", "--alluredir=./allure-results"]
            
            if markers:
                for marker in markers:
                    args.append(f"-m {marker}")

            args.extend(test_paths)
            pytest.main(args)
            return True
        except Exception as e:
            logger.error(f"测试执行失败: {str(e)}")
            return False

    @staticmethod
    @allure.step("设置测试环境")
    def setup_test_environment(env_config: Dict[str, Any]) -> None:
        """
        设置测试环境
        :param env_config: 环境配置
        """
        # 在这里实现环境设置逻辑
        pass

    @staticmethod
    @allure.step("清理测试环境")
    def cleanup_test_environment() -> None:
        """清理测试环境"""
        # 在这里实现环境清理逻辑
        pass 