requirements.txt
pytest>=7.4.0
pytest-xdist>=3.3.1
allure-pytest>=2.13.2
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
jsonpath>=0.82
loguru>=0.7.0
pymysql>=1.1.0 